Question_1/requirements.txt
praw==7.7.1
kafka-python==2.0.2
nltk==3.8.1
pyspark==3.5.0